% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceDerivedFeatures.R
\name{sequenceDerivedFeatures}
\alias{sequenceDerivedFeatures}
\title{Extract sequence predictive features from range-based data}
\usage{
sequenceDerivedFeatures(x, sequence, encoding = c("onehot", "iRNA"))
}
\arguments{
\item{x}{A \link{GRanges} object for the genomic ranges to be annotated, the \code{width} of x must all be equal.}

\item{sequence}{A \link{BSgenome} or \link{XStringSet} object for the genome sequence.}

\item{encoding}{Can be one of the following:
\describe{
\item{\strong{onehot}}{From the 5' end to the 3'end of x, 
each nucleotide position is coded by 4 indicators/dummy variables, 
where each dummy variable indicates that the position is equal to the base "A", "T", "C", and "G", respectively.}

\item{\strong{iRNA}}{
Each nucleotide position is encoded by 4 variables, 
the first variable indicates that the nucleotide is purine (A or G), 
the second variable indicates that the nucleotide has an amino group (A or C), 
and the third variable indicates the formation of weak hydrogen bond (A or T), 
the fourth variable is calculated by the cumulative frequency of nucleotides from 
the leftmost position to that position.
}
}}
}
\value{
A \code{data.frame} object whose number of rows is the length of \code{x}, and the number of columns is 4 times the width of \code{x}.
The column types in the \code{data.frame} are all numeric.
}
\description{
A function to extract sequence features from the input \code{\link{GRanges}} object and the \code{\link{BSgenome}} object.
}
\details{
The function first extract sequence within the genomic ranges defined by \code{x}. 
Then, the sequences are processed according to the selected encoding method.
}
\examples{
library(BSgenome.Hsapiens.UCSC.hg19)

## Define the Granges to be annotated:
set.seed(01)

X <- GRanges(rep(c("chr1", "chr2"), c(15, 15)),
             IRanges(c(sample(11874:12127, 15), sample(38814:41527,15)), width=5),
             strand=Rle(c("+", "-"), c(15, 15)))
   
## Extract onehot encoded sequence features
seq_onehot <- sequenceDerivedFeatures(X, Hsapiens, encoding = "onehot")       
str(seq_onehot)                

## Extract iRNA encoded sequence features
seq_iRNA <- sequenceDerivedFeatures(X, Hsapiens, encoding = "iRNA")
str(seq_iRNA)

}
\seealso{
\itemize{
\item{}{The \link{genomeDerivedFeatures} for extraction of genome-derived features.}
\item{}{The \link{topologyOnTranscripts} for calculation of the meta-tx topologies on transcripts of genes.}
}
}
