% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhistlerUsersGuide.R
\name{WhistlerUsersGuide}
\alias{WhistlerUsersGuide}
\title{View WhistleR User's Guide}
\usage{
WhistlerUsersGuide(view = TRUE)
}
\arguments{
\item{view}{logical, should the document be opened using the default PDF document reader?}
}
\value{
Character string giving the file location. If view=TRUE, the PDF document reader is started and the User's Guide is opened, as a side effect.
}
\description{
Finds the location of the WhistleR User's Guide and optionally opens it.
}
\details{
The function vignette("WhistleR") will find the short WhistleR Vignette which describes how to obtain the WhistleR User's Guide. The User's Guide is not itself a true vignette because it is not automatically generated using \link{Sweave} during the package build process. This means that it cannot be found using vignette, hence the need for this special function.
If the operating system is other than Windows, then the PDF viewer used is that given by Sys.getenv("R_PDFVIEWER"). The PDF viewer can be changed using Sys.putenv(R_PDFVIEWER=).
}
\examples{
\dontrun{
# To get the location:
WhistlerUsersGuide(view=FALSE)
# To open in pdf viewer:
WhistlerUsersGuide()
}
}
\seealso{
\link{system}
}
\author{
Zhen Wei
}
